Cleanup settings page(s)
- New separate directory
- Separate view for About information
